+++ contracts/interfaces/ICluster.sol
@@ -3,21 +3,21 @@ pragma solidity ^0.8.22;
 
 interface ICluster {
     function isWhitelisted(
-        uint16 lzChainId,
+        uint32 lzChainId,
         address _addr
     ) external view returns (bool);
 
     function updateContract(
-        uint16 lzChainId,
+        uint32 lzChainId,
         address _addr,
         bool _status
     ) external;
 
     function batchUpdateContracts(
-        uint16 _lzChainId,
+        uint32 _lzChainId,
         address[] memory _addresses,
         bool _status
     ) external;
 
-    function lzChainId() external view returns (uint16);
+    function lzChainId() external view returns (uint32);
 }
